planner refactor cont'd: vertices
