+++ compiler/annotation/type_seeder.rs
@@ -633,7 +633,7 @@ impl UnaryConstraint for Label<Variable> {
         )
         .map_err(|source| TypeInferenceError::ConceptRead { source })?;
         if let Some(annotation) = annotation_opt {
-            graph_vertices.add_or_intersect(self.left(), Cow::Owned(BTreeSet::from([annotation])));
+            graph_vertices.add_or_intersect(self.type_(), Cow::Owned(BTreeSet::from([annotation])));
             Ok(())
         } else {
             Err(TypeInferenceError::LabelNotResolved { name: self.type_label().to_string() })
@@ -660,7 +660,7 @@ impl UnaryConstraint for RoleName<Variable> {
                     .map_err(|source| TypeInferenceError::ConceptRead { source })?;
                 annotations.extend(subtypes.into_iter().map(|subtype| TypeAnnotation::RoleType(subtype.clone())));
             }
-            graph_vertices.add_or_intersect(self.left(), Cow::Owned(annotations));
+            graph_vertices.add_or_intersect(self.type_(), Cow::Owned(annotations));
             Ok(())
         } else {
             Err(TypeInferenceError::RoleNameNotResolved { name: self.name().to_string() })
